Additional-Push-AA

+++ DirtX.Core/Services/UserService.cs
@@ -1,6 +1,5 @@
 ﻿using DirtX.Core.Interfaces;
 using DirtX.Core.Models;
-using DirtX.Core.Models.Admin;
 using DirtX.Infrastructure.Data;
 using DirtX.Infrastructure.Data.Models.Users;
 using Microsoft.EntityFrameworkCore;
@@ -16,13 +15,37 @@ namespace DirtX.Core.Services
             context = _context;
         }
 
+        public async Task<EditProfileViewModel> GetUserByIdAsync(string userId)
+        {
+            if (!Guid.TryParse(userId, out Guid parsedUserId))
+            {
+                return null;
+            }
+
+            EditProfileViewModel userModel = await context.Users
+                .Where(u => u.Id == parsedUserId.ToString())
+                .Select(u => new EditProfileViewModel()
+                {
+                    Id = u.Id.ToString(),
+                    Email = u.Email,
+                    FirstName = u.FirstName,
+                    LastName = u.LastName,
+                    Country = u.Country,
+                    City = u.City,
+                    Address = u.Address,
+                })
+                .FirstOrDefaultAsync();
+
+            return userModel;
+        }
+
         public async Task<List<UserViewModel>> GetAllUsersAsync()
         {
             return await context.Users
                 .Select(u => new UserViewModel()
                 {
                     Id = u.Id.ToString(),
-                    Username = u.UserName,
+                    Email = u.Email,
                     IsAdmin = u.IsAdmin,
                     IsReseller = u.IsReseller,
                     CreatedOn = u.CreatedOn
@@ -30,11 +53,37 @@ namespace DirtX.Core.Services
                 .ToListAsync();
         }
 
+        public async Task<List<UserViewModel>> GetAllUsersWithoutCurrentAsync(string userId)
+        {
+            if (!Guid.TryParse(userId, out Guid parsedUserId))
+            {
+                return null;
+            }
+
+            List<UserViewModel> allUsers = await context.Users
+                .Where(u => u.Id != parsedUserId.ToString())
+                .Select(u => new UserViewModel()
+                {
+                    Id = u.Id.ToString(),
+                    Email = u.Email,
+                    IsAdmin = u.IsAdmin,
+                    CreatedOn = u.CreatedOn,
+                })
+                .ToListAsync();
+
+            return allUsers;
+        }
+
         public async Task EditProfileAsync(string userId, EditProfileViewModel model)
         {
+            if (!Guid.TryParse(userId, out Guid parsedUserId))
+            {
+                return;
+            }
+
             AppUser user = await context.Users
-                .Where(u => u.Id.ToString() == userId)
-                .FirstAsync();
+                .Where(u => u.Id == parsedUserId.ToString())
+                .FirstOrDefaultAsync();
 
             if (user != null)
             {
@@ -43,51 +92,42 @@ namespace DirtX.Core.Services
                 user.Country = model.Country;
                 user.City = model.City;
                 user.Address = model.Address;
-                user.ProfilePicture = model.ProfilePicture;
                 user.Email = model.Email;
 
                 await context.SaveChangesAsync();
             }
         }
 
-        public Task DeleteUserAsync(string userId)
+        public Task DeleteUser(string userId)
         {
+            if (!Guid.TryParse(userId, out Guid parsedUserId))
+            {
+                Exception ex = new("User not found!");
+                return Task.FromException(ex);
+            }
+
             AppUser user = context.Users
-                .Where(u => u.Id.ToString() == userId)
+                .Where(u => u.Id == parsedUserId.ToString())
                 .First();
 
-            //TODO - CHECK AGAIN IC CASE OF ERRORS, DELETE RELATED TABLES LIKE GARAGE TOO.
-
             context.Users.Remove(user);
-            context.SaveChangesAsync();
+            context.SaveChanges();
 
             return Task.CompletedTask;
         }
 
-        public async Task ToggleUserReseller(string userId)
-        {
-            AppUser user = await context.Users
-                .Where(u => u.Id.ToString() == userId)
-                .FirstAsync();
-
-            if (user.IsReseller is true)
-                user.IsReseller = false;
-            else
-                user.IsReseller = true;
-
-            await context.SaveChangesAsync();
-        }
-
         public async Task ToggleUserAdmin(string userId)
         {
+            if (!Guid.TryParse(userId, out Guid parsedUserId))
+            {
+                return;
+            }
+
             AppUser user = await context.Users
-                .Where(u => u.Id.ToString() == userId)
+                .Where(u => u.Id == parsedUserId.ToString())
                 .FirstAsync();
 
-            if (user.IsAdmin is true)
-                user.IsAdmin = false;
-            else
-                user.IsAdmin = true;
+            user.IsAdmin = !user.IsAdmin;
 
             await context.SaveChangesAsync();
         }
